# Backend/dist/openapi.yaml
openapi: 3.0.3
info:
  title: Grandma's recipes and others
  description: |- 
    API for recipes project
    This is a Server of Grandma's recipes and others based on the OpenAPI 3.0 specification.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
    
    # Introduction
    The currently supported resources are:

    * Preview details view
    * navigation bar
    * About page
    * Registration page Registration
    * login login page
    * Main recipe page
    * Recipe presentation page
    * Recipe search page
    * Modal recipe creation
    * My favorites page
    * My recipes page
    * My family recipes page
    * Bonus - recipe preparation page
    * Bonus - meal planning page

    
    This API will support the following HTTP actions on each resource:

    * GET - Retrieve a single resource or all resources in a collection.
    * POST - Create a new resource or collection of resources.

    In addition, this API will support Action resources that can be used to perform operations outside the standard CRUD operations.
    
    This API will be used to support common data connectors, such as Azure Data Connectors and Prismatic Data Connectors.
    

    Some Assignment links:
    - [First Assignment - Shay](https://web-development-environments-2023.github.io/315572677/)
    - [First Assignment - Oran](https://web-development-environments-2023.github.io/318872215/)
...
    
...
servers:
...
paths:
...
  /users/favorites:
    get:
      tags:
      - Users
      summary: Return the favorite recipes that were saved by user
      description: Get an array of recipes that were saved by a specific user
      responses:
        "200":
          description: data about the user's favorites recipes returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []

    post:
      tags:
      - Users
      summary: save recipe in the favorites list of a user
      description: save recipe in the favorites list of a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeId:
                  type: string
                  description: reciepe's id which the user saved
                  example: 716429
      responses:
        "201":
          description: The Recipe successfully saved as favorite
          $ref: '#/components/responses/Created'
        "401":
          description: user is not authorized
          $ref: '#/components/responses/Unauthorized'
      security:
      - cookieAuth: []

  # /users/favorites
  /users/lastWatched:
    get:
      tags:
        - Users
      summary: Get 3 last watched recipes by login user
      description: Return 3 last watched recipes by login user
      responses:
        '200':
          description: Successfully returned 3 last watched recipes
        '400':
          description: 'invalid input, object invalid'

  # /users/myRecipes
  /users/myRecipes:
    get:
      tags:
        - Users
      summary: Get all recipes user created
      description: Will include the list of all recipes created by the user.
      responses:
        '200':
          description: Successfully returned all recipes
        '401':
          description: 'user is not authorized'
        '404':
          description: 'user not found'
      security:
        - cookieAuth: []

  # /users/FamilyRecipes
  /users/familyRecipes:
    get:
      tags:
        - Users
      summary: Family recipes page view
      description: Will show your family recipes.
      responses:
        '200':
          description: LaFamilia page loaded successfully.
        '401':
          description: 'user is not authorized'
        '404':
          description: LaFamilia page not found.  

  # /users/Likes
  /users/Like:
    post:
      tags:
        - Users
      summary: Add like to recipe
      description: Add like for recipeID
      responses:
        '201':
          description: successful operation (add like to recipe in the DB).
        '400':
          description: Invalid ID of recipe supplied
        '401':
          description: User is not authorized
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Like' 

  # /users/myRecipes/{recipeId}
  /users/myRecipes/{recipeId}:
    get:
      tags:
      - Users
      summary: Return preview details on recipe by his ID
      description: Return preview details on recipe by his ID
      security:
        - ApiKeyAuth: [] 
      parameters:
      - name: recipeId
        in: path
        description: recipe's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: data about the recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePreview'
        '401':
          description: User is not authorized
        "404":
          description: recipeId is not exists
          $ref: '#/components/responses/NotFound'

  # /users/checkWatched/:recipeId
  /users/checkWatched/{recipeId}:
    get:
      tags:
        - Users
      summary: Get True\False if recipe watched by login user
      description: Return True\False for watched recipe by login user
      responses:
        '200':
          description: Successfully true\false if the recipe exists
        '400':
          description: 'invalid input, object invalid'
        '401':
          description: User is not authorized

  # /users/checkWatched/:recipeId
  /users/favorites/{recipeId}:
    get:
      tags:
      - Users
      summary: Return True\False favorite recipes that were saved by user
      description: Get an True\False of recipes that were saved by a specific user
      responses:
        "200":
          description: data about the user's favorites recipes returned
          content:
            application/json:
              schema:
                type: boolean
                items:
                  $ref: '#/components/schemas/RecipePreview'
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []

  # /users/Like/{recipeId}
  /users/Like/{recipeId}:
    get:
      tags:
        - Users
      summary: Get True\False if recipe liked by login user
      description: Return True\False for liked recipe by login user
      responses:
        '200':
          description: Successfully true\false if the recipe exists
        '400':
          description: 'invalid input, object invalid'
        '401':
          description: User is not authorized
      
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
# End of users routes



#~~~~~~~~~~~~~ recipes ~~~~~~~~~~~~~# 
  # recipes/{recipeId}
...
components:
...
  responses:
...
    Created:
      description: Created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Unauthorized:
      description: Authorization information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 0
        message:
          type: string
          example: "message depend on status code"
      required:
        - status
        - message

    # Register
...
    RecipePreview:
      type: object
      properties:

        id:
          type: integer
          description: recipe identifier
          example: 102

        title:
          type: string
          description: recipe name
          example: Simple Macaroni and Cheese

        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeimages/655455-566x370.jpg

        readyInMinutes:
          type: integer
          description: the time takes to make this recipe
          example: 45

        popularity:
          type: integer
          description: how many likes got this recipe
          example: 5987

        vegetarian:
          type: boolean
          description: indication if the recipe is vegetarian
          example: false

        vegan:
          type: boolean
          description: indication if the recipe is vegan
          example: false

        glutenFree:
          type: boolean
          description: indication if the recipe is gluten free
          example: false
  
    # Search
...
    Like:
      type: object
      required:
        - recipeId
      properties:

        recipeId:
          type: integer
          description: amount of likes
          example: 55



    # AnalyzedRecipe
...